wrongtype wave 3: 35 per-signature retypes + caller reconciliation
Genuine pointer/width/ref/enum fixes to match the binary, then chased the caller
cascade without casts (net -14 casts across the diff; dev build green):
- void* -> Player*/AEGeometry*/LodMeshMerger*/bool* (NewsItem, PlayerFighter, AEGeometry, ...)
- Array<Item*>* -> Array<int>* (Item ctor; callers retyped their locals)
- int -> enum (setRotationOrder -> RotationOrder; callers pass ROTATION_ORDER_YXZ)
- String*/String const& -> String by-value (Agent, Station, Wanted, PendingProduct, Mission, ...)
- Vector const* -> Vector const& (Explosion start/translate/update; callers deref/switch overload)
- unsigned int -> int width (Standing, TextureAlphaTestShader), int -> bool (Level, ImageFactory)
Reconciled 17 caller TUs by removing now-wrong casts ((unsigned)/(char)/(int)(intptr_t))
and adjusting args (&x -> x, retyped locals), never adding casts.

Deferred to dedicated passes (reverted here to keep the wave clean): RenderEffect
virtual family (base+5 shaders), FModSound events[] member retype, IParticleSystem
+ PaintCanvas relocation. Flagged (need field retype / reverse-namespace): Mission
ctor x2, AEGeometry::DEBUG_setMeshMergerIndex, BoundingVolume (recheck).

+++ src/game/menu/MGame.cpp
@@ -717,7 +717,7 @@ void MGame::OnTouchBegin(int p1, int p2, void *touchId) {
             int ad = ApplicationManager_GetApplicationData();
             if (*(uint8_t *)((char *)ad + 5) != 0) return;
             if (*(uint8_t *)((char *)ad + 0xc) != 0) return;
-            self->menuWindow->OnTouchBegin(p1, p2, (int)(intptr_t)touchId);
+            self->menuWindow->OnTouchBegin(p1, p2, touchId);
             if (self->freeCamMode == 0) return;
             if (self->menuWindow->isShowingMessage() != 0) return;
             if (self->menuWindow->isMakingScreenshot() != 0) return;
